Cambios para proyecto final

+++ src/controllers/session.controllers.js
@@ -1,15 +1,16 @@
-//import UserManager from "../daos/mongodb/daos/UserManager.class.js";
 import SessionService from "../services/session.service.js";
 import { createHash, validatePassword} from "../../src/utils.js";
 import jwt from 'jsonwebtoken' 
 import config from "../config.js"; 
 import Mail from "../helpers/mail.js";
+import CartService from "../services/carts.service.js";
 
 export default class SessionController{ 
 
   
   constructor(){
     this.sessionService=new SessionService()
+    this.cartService=new CartService();
     this.mail = new Mail()
   }
 
@@ -31,9 +32,8 @@ async loginController(req,res){
    role: req.user.role,
    id: req.user._id,
    cart: req.user.cart
-   } 
+   }  
 
-   
    let token = jwt.sign(user, config.JWT_SECRET, {
     expiresIn: "24h",
   });
@@ -43,19 +43,17 @@ async loginController(req,res){
   await this.sessionService.updateUserLastConnection(req.user._id);
  } 
 
-async loginFailController(req, res) {
- console.log("Fallo la autenticación del login");
+async loginFailController(req, res) { 
  res.status(400).send({status: "error", error: "Fallo la autenticación del login"})
 }
  
 async registerFailController(res) {
-  console.log("Fallo la autenticación");
+ 
   res.status(400).send({status: "error", error: "Fallo la autenticación"})
 }
 
 async getCurrentController(req, res) {
     const result=await this.sessionService.getCurrentService(req, res);
-    //res.send(req.user);
     res.send(result);
 }
 
@@ -90,11 +88,10 @@ async githubcallbackController(req, res){
 async resetPasswordController(req,res){
 
   const {email,password} = req.body;
-console.log(email, password)
+ 
  if(!email||!password) return res.status(400).send({status:"error",error:"No completo los campos del Formulario"});
   const user = await this.sessionService.getUserByIdService(email);
-  
-  //await managerUsers.getUserById(email);
+ 
   if(!user) return res.status(404).send({status:"error",error:"El usuario no se encuentra registrado"});
   
   try{
@@ -112,18 +109,13 @@ console.log(email, password)
 
 //envia correo - email
   async requestResetPasswordController(req,res){
-
     const {email} = req.body;
-    console.log(email)
+
     if(!email) return res.status(400).send({status:"error",error:"No completo los campos del Formulario"});
     const user = await this.sessionService.getUserByIdService(email);
-    
-    //await managerUsers.getUserById(email);
     if(!user) return res.status(404).send({status:"error",error:"El usuario no se encuentra registrado"});
     
-    try{
-      //console.log("Entro a mandar el emailssssssssss");
-      
+    try{ 
       let tokenReset = jwt.sign({email}, config.TOKEN_RESET, {
         expiresIn: "1h",
       });
@@ -135,8 +127,7 @@ console.log(email, password)
       http://localhost:8080/resetPassword?token=${tokenReset}</div>`);
      let asunto="Correo de Recuperación de Contraseña";
      const result=this.sessionService.enviarCorreo(email,asunto,html,res);
-   //  const result = this.mail.send(email, "Correo de Recuperación de Contraseña", html);
-      return result; 
+      return res.status(200).send({status: "success", result: result});
     }catch(error){
       return res.status(404).send({status: "error", error: error.message});
     }
@@ -145,25 +136,24 @@ console.log(email, password)
   async cambiarRolController(req,res){
     try {
       let id = req.params.uid
-      let user = await this.sessionService.getUserService(id);
-
+      let user = await this.sessionService.getUserService(id); 
       if(user.role=="user"){
         //Agrego control para saber si el usuario tiene los archivos necesarios para ser Premium
         const result=await this.sessionService.getPremiumRequiredDoc(id);
         if(result){
           await this.sessionService.updateUserRoleService(id, "premium");
-          return res.send({status: "success", message: "Rol de Usuario actualizado a Premium"});
+          return res.status(200).send({status: "success", message: "Rol de Usuario actualizado a Premium"});
         }else{
           return res.status(400).
-          send({status: "failure", details: "El usuario no tiene la documentación requerida para pasar a Premium."})
+          send({status: "failure", message: "El usuario no tiene la documentación requerida para pasar a Premium."})
 
         }
       }else if(user.role=="premium"){
         await this.sessionService.updateUserRoleService(id, "user");
-        return res.send({status: "success", message: "Rol de Premium actualizado a Usuario"});
+        return res.status(200).send({status: "success", message: "Rol de Premium actualizado a Usuario"});
       }else{
         return res.status(400).
-        send({status: "failure", details: "No puede cambiar de Rol."})
+        send({status: "failure", message: "No puede cambiar de Rol."})
       }
   
     }
@@ -212,7 +202,6 @@ async updatePathDocuments(req,res){
   }
 }
 
-
 //Controlo si tiene la documentación requerida para ser Premium
 async getPremiumRequiredDoc(req,res){
   try {
@@ -237,7 +226,8 @@ async getPremiumRequiredDoc(req,res){
 
 async getUsersController(req, res) {
   const result=await this.sessionService.getUsersService();
-  res.send(result);
+  res.status(200).send(result)
+  
 }
 
 async deleteUsersInactivosController(req, res) {
@@ -249,18 +239,35 @@ async deleteUsersInactivosController(req, res) {
       `<div><h1>Se procede a eliminar su usuario del Sistema, ya que hemos detectado que no tiene actividad reciente.</h1> 
          Ante cualquier inquietud contáctenos! </div>`);
 
+    if(usuarios.length==0) return res.status(404).send({status: "error", message: "No se encontraron usuarios inactivos!"});
+
     usuarios.forEach((user)=>{
        
       let asunto=`Usuario  - ${user.email}`;
       this.sessionService.enviarCorreo(user.email,asunto,cuerpo)
       this.sessionService.deleteUserInactivo(user);
+      this.cartService.deleteCartService(user.cart);      
     })
-    
+    return res.status(200).send({status: "success", message: "Usuarios Eliminados con Exito!"});
   }catch(error){
     return res.status(404).send({status: "error", error: error.message});
   }
 
-  res.status(200).send({status: "success", message: "Usuarios Eliminados con Exito!"});
+  
 }
- 
+
+
+async deleteUserController(req, res) {  
+    try{ 
+      let id = req.params.uid;       
+      let user = await this.sessionService.getUserService(id); 
+      this.cartService.deleteCartService(user.cart);
+      this.sessionService.deleteUser(id);
+      
+      res.status(200).send({status: "success", message: "Se ha eliminado el usuario con éxito"});
+    }catch(error){
+      return res.status(404).send({status: "error", error: error.message});
+    }
+
+  }
 }
